fix(tests): repair the ctest tier and de-collide unity-build file statics
Wave-7 TU splits left several unit tests unbuildable: three CSV suites lost their
QTEST_APPLESS_MAIN entry points, and link sets drifted (LanguageTable/Translator coupling,
Sparkplug suites reaching QColor through OpcUaWire.h, table-snapshot-channel pulling frame
serialization, Modbus generation needing BUILD_COMMERCIAL). File-local statics that now share
a unity batch (projectModel(), kTerminateGraceMs, kBackendName, ...) get file-specific names.

Also: appendArcTo probes elementCount() instead of isEmpty(), since a lone-moveTo path reports
empty but must still seed the tangent solve; the terminal wrap test narrows columns after the
append, matching how wrapped rows actually come to exceed the live width; enum QCOMPAREs cast
to int so QTest needn't format the enum class.

+++ app/tests/tst_widget_manifest.cpp
@@ -184,7 +184,8 @@ void TstWidgetManifest::acceptsWellFormedPackage()
   QVERIFY(result.descriptor.isValid());
   QCOMPARE(result.descriptor.id, QStringLiteral("com.example.dial"));
   QCOMPARE(result.descriptor.qmlFile, QStringLiteral("Widget.qml"));
-  QCOMPARE(result.descriptor.scope, UI::WidgetExtensions::GroupScope);
+  QCOMPARE(static_cast<int>(result.descriptor.scope),
+           static_cast<int>(UI::WidgetExtensions::GroupScope));
   QCOMPARE(result.descriptor.readsStringValues, true);
   QCOMPARE(result.descriptor.bundled, false);
   QVERIFY(result.descriptor.replaces.isEmpty());
@@ -219,7 +220,8 @@ void TstWidgetManifest::clampsDeclaredBounds()
   QVERIFY(result.ok);
   QCOMPARE(result.descriptor.accepts.minDatasets, 4);
   QCOMPARE(result.descriptor.accepts.maxDatasets, 4);
-  QCOMPARE(result.descriptor.accepts.value, UI::WidgetExtensions::StringValue);
+  QCOMPARE(static_cast<int>(result.descriptor.accepts.value),
+           static_cast<int>(UI::WidgetExtensions::StringValue));
   QCOMPARE(result.descriptor.defaultWidth, 48);
   QCOMPARE(result.descriptor.defaultHeight, 8192);
 }
